+++ src/app/(main)/(mobile)/me/(home)/__tests__/UserBanner.test.tsx
@@ -24,6 +24,10 @@ vi.mock('@/features/User/UserLoginOrSignup/Community', () => ({
   default: vi.fn(() => <div>Mocked UserLoginOrSignup</div>),
 }));
 
+vi.mock('@/const/version', () => ({
+  isDeprecatedEdition: false,
+}));
+
 // 定义一个变量来存储 enableAuth 的值
 let enableAuth = true;
 let enableClerk = false;