+++ src/types/session/index.ts
@@ -15,3 +15,11 @@ export interface UpdateSessionParams {
   pinned?: boolean;
   updatedAt: Date;
 }
+
+export interface SessionRankItem {
+  avatar: string | null;
+  backgroundColor: string | null;
+  count: number;
+  id: string;
+  title: string | null;
+}